+++ packages/spacecat-shared-data-access/test/fixtures/index.fixtures.js
@@ -29,6 +29,7 @@ import sites from './sites.fixture.js';
 import suggestions from './suggestions.fixture.js';
 import fixEntities from './fix-entity.fixture.js';
 import pageIntents from './page-intents.fixture.js';
+import reports from './reports.fixture.js';
 
 export default {
   apiKeys,
@@ -50,4 +51,5 @@ export default {
   scrapeJobs,
   scrapeUrls,
   pageIntents,
+  reports,
 };